Flip the use of triangle icon on the right of settings.
Triangle will now only show up for preferences that have dialogs popup
when selected - like lists, sliders, EditTexts and Yes/No dialogs.
This is a change for the base preferences. If any subclasses override
and specify a different layout or normal preferences display a
dialog when pressed, they may not follow the new style.