Fixing a concurrency problem by changing the contact aggregation exception API.
The current API requires a contact_id and a raw_contact_id
There are at least two issues with this approach I did not recognize initially:

1. Contact_id may be changed asynchronously by aggregation or some other process.
2. A raw contacts may need to be added to an aggregate before the actual aggregation pass
has gotten to it, so the client would need to wait for the aggregation to complete
before it can set an aggregation exception.  That's backwards.