Don't scroll the title bar off screen while loading.
Though the user can still manually scroll the title bar
off screen, we do not want it to happen automatically.
When we load a new page, its scroll position can be
set to (0,0).  Make sure that we simply set it to the
top without adding in the title bar's height.  Also,
prevent javascript which attempts to scroll the page
to (0,0) or (0,1) from removing the title bar.

Fixes http://b/issue?id=2113398 and http://b/issue?id=2123079

Change-Id: Ida64d8c94be6744c7b0b4d60af1c229b2cc83673