Moving DISPLAY_NAME to column "data1" to keep it "aligned" with other data types.
This is a minor optimization for the sake of the aggregator.  The aggregator now relies on the display name
instead of structured components. As a result, we only need one column from the data table
for the structured name.  For other data types it's data1. Might as well use the same column
for display name.

Change-Id: Ib22d1f1a7a91f12716d1a460e1578f01926c393a