set the selection after sending a click
Webkit expects the currently focused text input to have
a set selection before it can receive key events for insertion.

Coincidentally, we always set the selection except when
using the 'next' button on the IME to switch between text fields.

The selection has to be set after the text field receives
focus, so add a setSelection() message after the click() message
is sent in passToJavaScript().

fixes http://b/issue?id=2087546