Remove keyguard view after a delay instead of just hiding it.
The delay should make sure the animation works, and by still removing it, we
preserve the initialization logic that depends on being reinstantiated each time
the screen turns off.  The logic could be adjusted to work without this, but
seems safer to maintain the existing behavior. Also reduces risk from having
keyguard around all the time.