Fix issue #2161169: onStop not being called.
There was a bug with the starting window where it could be added to
the app window list twice, so the buddy list would end up with one
left over after all was done.  This would result in visibility
changes not being delivered to it correctly, delaying the dispatch
of onStop.

Change-Id: If1993eaf9cfbba1f523ce5aaa478be0239d0c7db