Fix issue #2179931: Power key to wake up frequently ignored when in dock app
The dock app is forcing the screen to a particular brightness level.  This
causes the window manager to often call into the power manager with the
new brightness.  This causes us to go in to updateLightsLocked() to figure
out and apply the real brightness to use.  When the screen is off the
real brightness always remains 0, but even if it didn't change from the
last one we would start an animation which would when done put the system
to sleep and fight with the user trying to turn the device on.

Now, if the new target brightness is the same as the last one, we leave the
animation as-is -- either running or not as appropriate.

Change-Id: I067d55ea2b39e294c5d5291587a4d8727c0b8083